Add initial files.
This populates the repo with a very minimal set of files. This is just
enough to set out a basic outline. Hopefully as we move forward we'll
be replacing many of these parts, but this is just something basic to
get things started.

This tree isn't really usable yet, as it doesn't yet have an easy way
to obtain a compiler-rt/libgcc build.