Rename `poll` to `poll_oneoff`, and remove `__wasi_subrwflags_t`.
Renaming `poll` avoids a conflict with the libc poll in implementations
where libc shared a namespace with the wasi syscall names. And calling
it "oneoff" makes it clear that this is the classic poll which unlike
things like epoll doesn't have a way to factor out the
per-file-descriptor overhead over multiple calls.

__wasi_subrwflags_t was unused and its only value was deprecated.