Remove `file_create` and add `file_mkdir`.
Besides being a slight simplification (the only current use of
`file_create` is creating directories), if we add ways to create
other kinds of filesystem objects in the future, it's desirable
to create separate functions for them.

WASI has two levels of capabilities -- static and dynamic. The
static capabilities are the imports. By using separate functions
for separate file types, we can distinguish between the
capabilities to create different types of filesystem objects
statically.

And, users won't often need to specify what type of thing to create
dynamicaly. For example, the POSIX interface just has `mkdir`.