Remove fd_dup and change fd_replace to fd_renumber.
This avoids committing core WASI to the "file descriptor" vs "file
description" distinction. And `dup` is mainly useful for (a) things that
really ought to be done in other ways, and (b) manipulating file
descriptors in the proximity of a fork, and wasi doesn't have fork.

This also removes the emulated mmap's support for PROT_WRITE, since it
depened on dup, however this wasn't very useful.