build: restructure the commands, remove options
This restructures the sysroot generation to reduce duplication by merging
the commands that write to the same location.  This should also speed up
the job a bit by invoking fewer commands.

Remove a few options that are not needed.

This now no longer has an always out of date target allowing null-builds.