Try to restore some build system conventions in a low-impact way
Fixes #269

The "don't move the variables independently if you want to `make
install`" part is a gross, but as neither `wasi-sdk` or Nixpkgs actually
needs `make install` (we both can just build right into the directories
we want) I figured it wasn't actually so heinous in practice.

Also reorganize the the github actions CI slightly.