review: avoid waiting for the TID to be written
Previously, the TID was only passed to the child thread entry point; the
parent thread was forced to wait until the child thread set the TID in
the pthread structure. With this change, the TID will be passed not only
to the child thread but also returned to the parent thread, so that
either side can make progress. The `i32.store` becomes an
`i32.atomic.store` to avoid concurrent writes; perhaps the entire field
should be made `atomic_int` to prevent subsequent reads from interfering
(?).