Default wasm32-wasip2 objects to `-fPIC` by default
This commit updates all of the `*.a` archives built for the
`wasm32-wasip2` target to use `-fPIC` instead of the default non-PIC
flag. This makes them suitable for linking into either a dynamic library
or a non-dynamic binary without recompiling libc. This is intended to
help integrate shared libraries into other precompiled languages, such
as Rust's standard library, as it makes its way upstream.

Lots of discussion about this also happened on #429 so I wanted to
highlight that this is only happening for the wasm32-wasip2 target and
no other targets. For example neither wasm32-wasi nor wasm32-wasip1 are
affected. Only users of wasm32-wasip2, of which there aren't too too
many, are possibly affected by the runtime overheads here. I have not
measured the hypothetical runtime myself, however.