Properly stub out pthread cancellation
Although it is not supported in WASI, thread cancellation is an important
feature of pthreads with wide-ranging implications.

Provide the best stubs we can within the limitations by allowing
cancellation type and state to be set, even though they will have no
effect as pthread_cancel is stubbed to always return ENOTSUP

Remove the THREAD_MODEL=single guard as it's currently not being
compiled in that case, and because we intend to provide
single-threaded stub implementations in the future.
Incidentally replace the existing pthread_setcancelstate stub
with one which actually correctly writes a value to the old parameter.