Parse service name as port number in getaddrinfo
Previously, the `getaddrinfo` function did not resolve service names into port
numbers, resulting in `sin_port` and `sin6_port` being hard-coded to `0`.

This commit adds support for `getaddrinfo` to parse the service string as
a port number if it is provided in numeric form. However, this commit does not
include support for resolving service names into port numbers, as the
getservbyname() function has not been implemented (yet). In this case
the `EAI_NONAME` error is returned.