Add `version.h` defining `WASI_LIBC_VERSION`
This change re-imagines [wasi-sdk#487] entirely in `wasi-libc`. Instead
of generating a `version.h` header late--in `wasi-sdk`--it is generated
immediately when `wasi-libc` is built. This has the disadvantage that we
don't know what `wasi-sdk` this will be used in... but almost: when
releasing `wasi-sdk`, we _should_ be tagging the `wasi-libc` repository.
If we did that more faithfully ([docs]), this commit would generate:

  ```c
  // Generated by wasi-lib's `Makefile`.
  #ifndef VERSION_H
  #define VERSION_H
  #define WASI_LIBC_VERSION "wasi-sdk-24"
  #endif
  ```

Using the magic of `git-describe`, if someone is using `wasi-libc`
directly, they may see a version like: `wasi-sdk-22-19-g5d3c5e9-dirty`.
This should be read like: `<wasi-sdk tag>-<# commits since
tag>-<current commit>-<is the repository dirty?>`.

[wasi-sdk#487]: https://github.com/WebAssembly/wasi-sdk/pull/487
[docs]: https://github.com/WebAssembly/wasi-sdk/blob/754aec3/RELEASING.md?plain=1#L22